==> 1_load_QC_preprocess/2_label_merge.rmd <==
---
title: "scRNAseq Neonatal Lung Injury 1.2"
author: Shawyon Shirazi & Nick Negretti
date: 3/14/24
output: rmarkdown::github_document
---

## Load libraries and helper functions
```{r, results="hide", message = FALSE, load-libraries}
source("./helper_functions/globals.R")
source("./helper_functions/libraries.R")
source("./helper_functions/cluster.R")
source("./helper_functions/colors.R")

N_WORKERS <- 12
plan("multicore", workers = N_WORKERS)
```

# Load premerged object
```{r load-data}
alldata <- readRDS("./data/checkpoint/load_all_data.rds")
```

# Function to merge list of data & split project name into descriptive labels
```{r merge}
merge_data <- function(alldata) {
  alldata <- Reduce(function(x, y) {merge(x, y)}, alldata)
  alldata$dataset <- "Unassigned"
  alldata$age <- "Unassigned"
  alldata$condition <- "Unassigned"
  alldata$broad_condition <- "Unassigned"
  alldata$dataset <- sapply(str_split(alldata$orig.ident, "_"),
                              function(x) {x[1]})
  alldata$age <- sapply(str_split(alldata$orig.ident, "_"),
                          function(x) {x[2]})
  alldata$condition <- sapply(str_split(alldata$orig.ident, "_"),
                                function(x) {x[3]})
  alldata$broad_condition <- sapply(str_split(alldata$orig.ident, "_"),
                                      function(x) {x[4]})
  return(alldata)
}
```

# Merge the data and save for next step
```{r merge all}
alldata <- merge_data(alldata)

saveRDS(alldata, "./data/checkpoint/label_merge.rds", compress = "bzip2")
```
